ci: drive mypy strict to zero errors
17 strict-mypy errors had accumulated across the audit cluster
merges.  All resolved:

  Provider attribute access:
    * `bench/_cli.py` — `Provider` protocol doesn't declare `chat` /
      `embedder`; switched to `getattr(provider, ..., None)` since
      both attrs are implementation-detail.  None-guard before stash.

  Retry signature variance:
    * `providers/openai.py` + `anthropic.py` — `Retry.call` /
      `Retry.acall` are `Callable[P, T]` / `Callable[P, Awaitable[T]]`
      via `ParamSpec`; the OpenAI / Anthropic SDK methods don't
      reduce cleanly under that parameterization.  Added
      `# type: ignore[arg-type]` at the call sites with the runtime
      contract documented.

  Async surface:
    * `integrations/_agent.py::achat` — `chat` takes `history` as a
      keyword-only arg; the prior `asyncio.to_thread(self.chat,
      user_message, history)` passed it positionally.  Switched to
      `history=history` so the kwarg-only signature is respected and
      the type-checker sees correct dispatch.

  Reachability:
    * `_otel.py::_Metrics._init`, `consolidation/_clustering.py::
      _hdbscan_available`, `storage/sqlite.py::initialize` — all had
      double-checked-locking patterns that mypy flagged as
      "unreachable" because it couldn't prove a concurrent caller
      could mutate the guarded flag.  Refactored each to the
      equivalent positive form (`if not self._initialized: ...`)
      which preserves the race semantics and silences the unreachable
      check.

  Return-type tightening:
    * `_vec_math.normalize`, `retrieve/_engine._get_created_at_batch`,
      `consolidation/_engine._matches_tenant` — explicit local
      annotations / `bool()` cast to satisfy "Returning Any from
      function declared to return X".

  Generic args:
    * `retrieve/_engine.py` — `_Counter[str | None]` for the per-
      session counter (was bare `_Counter`).

  Cleanup:
    * `_mmr.py` — re-cast post-normalize matrix to `float32` so the
      assignment matches the declared dtype.
    * `decay/_engine.py` — removed unused `type: ignore[assignment]`
      on the clock callable.

Plus the lint fix in the previous commit closes the ruff side.
Three CI gates (`ruff check`, `ruff format --check`, `mypy`) now
pass locally; 1217 tests still pass.

+++ src/engram/retrieve/_mmr.py
@@ -110,7 +110,7 @@ def mmr_select(
     norms = np.linalg.norm(matrix, axis=1, keepdims=True)
     # Avoid division by zero: zero-norm rows stay zero after divide.
     norms_safe = np.where(norms > 0, norms, 1.0)
-    matrix = matrix / norms_safe
+    matrix = (matrix / norms_safe).astype(np.float32)
 
     remaining = np.ones(n, dtype=bool)
     redundancy = np.zeros(n, dtype=np.float32)
